+++ modules/features/sharing/src/main/kotlin/au/com/shiftyjelly/pocketcasts/sharing/timestamp/ShareEpisodeTimestampPage.kt
@@ -9,10 +9,12 @@ import androidx.compose.ui.tooling.preview.Preview
 import au.com.shiftyjelly.pocketcasts.models.entity.Podcast
 import au.com.shiftyjelly.pocketcasts.models.entity.PodcastEpisode
 import au.com.shiftyjelly.pocketcasts.sharing.social.SocialPlatform
+import au.com.shiftyjelly.pocketcasts.sharing.ui.CardType
 import au.com.shiftyjelly.pocketcasts.sharing.ui.Devices
 import au.com.shiftyjelly.pocketcasts.sharing.ui.HorizontalEpisodeCard
 import au.com.shiftyjelly.pocketcasts.sharing.ui.HorizontalSharePage
 import au.com.shiftyjelly.pocketcasts.sharing.ui.ShareColors
+import au.com.shiftyjelly.pocketcasts.sharing.ui.SquareEpisodeCard
 import au.com.shiftyjelly.pocketcasts.sharing.ui.VerticalEpisodeCard
 import au.com.shiftyjelly.pocketcasts.sharing.ui.VerticalSharePage
 import com.airbnb.android.showkase.annotation.ShowkaseComposable
@@ -86,14 +88,31 @@ private fun VerticalShareEpisodeTimestampPage(
             listener.onShare(podcast, episode, timestamp, platfrom)
         }
     },
-    middleContent = {
+    middleContent = { cardType, modifier ->
         if (podcast != null && episode != null) {
-            VerticalEpisodeCard(
-                podcast = podcast,
-                episode = episode,
-                useEpisodeArtwork = useEpisodeArtwork,
-                shareColors = shareColors,
-            )
+            when (cardType) {
+                CardType.Vertical -> VerticalEpisodeCard(
+                    podcast = podcast,
+                    episode = episode,
+                    useEpisodeArtwork = useEpisodeArtwork,
+                    shareColors = shareColors,
+                    modifier = modifier,
+                )
+                CardType.Horiozntal -> HorizontalEpisodeCard(
+                    podcast = podcast,
+                    episode = episode,
+                    useEpisodeArtwork = useEpisodeArtwork,
+                    shareColors = shareColors,
+                    modifier = modifier,
+                )
+                CardType.Square -> SquareEpisodeCard(
+                    podcast = podcast,
+                    episode = episode,
+                    useEpisodeArtwork = useEpisodeArtwork,
+                    shareColors = shareColors,
+                    modifier = modifier,
+                )
+            }
         }
     },
 )
