+++ modules/features/sharing/src/main/kotlin/au/com/shiftyjelly/pocketcasts/sharing/ui/CardType.kt
@@ -1,0 +1,7 @@
+package au.com.shiftyjelly.pocketcasts.sharing.ui
+
+internal enum class CardType {
+    Vertical,
+    Horiozntal,
+    Square,
+}
